refactor: move role selection form into view component
