+++ src/__tests__/pages/onboarding/school-selection.test.tsx
@@ -7,7 +7,7 @@ import SchoolSelection from "@/pages/onboarding/school-selection";
 jest.mock("posthog-js/react", () => ({
   useFeatureFlagEnabled: () => true,
 }));
-describe("onboarding page", () => {
+describe("Onboarding school selection page", () => {
   test("it renders the onboarding page", () => {
     renderWithProviders()(<SchoolSelection />);
